Change structures of contract codes
