+++ answer/AppleStamp.java
@@ -1,6 +1,6 @@
 package com.r3.developers.apples.states;
 
-import com.r3.developers.apples.contracts.AppleStampContract;
+import com.r3.developers.apples.contracts.stamp.AppleStampContract;
 import net.corda.v5.base.annotations.ConstructorForDeserialization;
 import net.corda.v5.ledger.utxo.BelongsToContract;
 import net.corda.v5.ledger.utxo.ContractState;
